ISSUE: [#] PURPOSE: Misc. cleanup for Maureen DESCRIPTION:
A lot of fairly mechanical code cleanups. Two big goals: (1) Track down old calls to the localizer, and replace them with unlocalized logging calls, as appropriate. (2) Get rid of the yellow flags in Eclipse. (For a given file, once all the localizer calls are gone, then the localizer is unused and we get the warning flag... and it's easier to spot this if there aren't other warnings.)

Types of edits made:
  on unused code and assignments --> // TODO: unused
    (These sometimes cascade: a variable is only set up to be used somewhere else.)
  on unreachable code --> // TODO: no-op
  on localized messages that used to go to the UI --> log.info("old reference: ...");
  some types parameterized, as appropriate
  on others --> @SuppressWarning("rawtypes")
  on some things that REALLY do need to be sorted out: @SuppressWarnings("null");